+++ src/Parser.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TupleSections #-}
 
 module Parser where
 
@@ -9,8 +10,6 @@ import Data.Text (Text, pack)
 import Text.Megaparsec hiding (State)
 import Text.Megaparsec.Char (char, string', alphaNumChar)
 import Data.Map.Ordered as O (fromList)
-import Text.Megaparsec.Debug (dbg)
-
 
 parse :: Text -> Either Text Q.Query
 parse query =
@@ -29,28 +28,27 @@ columns = O.fromList <$> columns'
 
 columns' :: Parser [(Q.Col, Ex)]
 columns' = do
-  headCol <- column -- <|> asColumn
+  headCol <- column
   tailCols <- (lex0 (char ',') >> columns') <|> return empty
   return $ headCol : tailCols
 
 column :: Parser (Q.Col, Ex)
-column = do
-  ex <- parseEx
-  case ex of
-    (Var name) -> return (name, ex)
-    _ -> asColumn ex
+column =  try (parseEx >>= asColumn) <|> ((\n -> (n, Var n)) <$> colName)
 
 asColumn :: Ex -> Parser (Q.Col, Ex)
-asColumn ex = do
-  _ <- lex1 (string' "AS")
-  name <- lex0 $ some $ alphaNumChar <|> char '_'
-  return (pack name, ex)
+asColumn ex = fmap (, ex) asName
+
+asName :: Parser Q.Col
+asName = lex1 (string' "AS") *> colName
+
+colName :: Parser Q.Col
+colName = pack <$> lex0 (some (alphaNumChar <|> char '_'))
 
 pFrom :: Parser Q.From
-pFrom = Q.From <$> (lex1 (string' "FROM") *> lex1 tableName)
+pFrom = Q.From <$> (lex1 (string' "FROM") *> tableName)
 
 tableName :: Parser Q.Table
-tableName = pack <$> some (alphaNumChar <|> char '_' <|> char '-' <|> char '/')
+tableName = pack <$> lex1 (some (alphaNumChar <|> char '_' <|> char '-' <|> char '/'))
 
 pWhere :: Parser Q.Where
 pWhere = Q.Where <$> (lex1 (string' "WHERE") *> parseEx)