+++ test/fixtures/queries/select_type_fail_cat.sql
@@ -1,0 +1,2 @@
+SELECT age ++ 3 AS concatInts
+FROM test/fixtures/tables/people