+++ test/fixtures/queries/select_type_fail_multi.sql
@@ -1,0 +1,2 @@
+SELECT "Hello " ++ True AS hiTrue, 4 = "world" as numWorld
+FROM test/fixtures/tables/people