+++ test/fixtures/queries/select_type_fail_not.sql
@@ -1,0 +1,2 @@
+SELECT NOT age AS notAge
+FROM test/fixtures/tables/people